Move deprecated decorator to a Decorators file
This is a better location as we can collect several decorators there.
